fix collectionID in SIOBlock

+++ include/podio/SIOBlock.h
@@ -104,7 +104,7 @@ public:
 
   SIOCollectionIDTableBlock(podio::EventStore* store);
 
-  SIOCollectionIDTableBlock(std::vector<std::string>&& names, std::vector<int>&& ids, std::vector<std::string>&& types,
+  SIOCollectionIDTableBlock(std::vector<std::string>&& names, std::vector<uint64_t>&& ids, std::vector<std::string>&& types,
                             std::vector<short>&& isSubsetColl) :
       sio::block("CollectionIDs", sio::version::encode_version(0, 3)),
       _names(std::move(names)),
@@ -131,7 +131,7 @@ public:
 
 private:
   std::vector<std::string> _names{};
-  std::vector<int> _ids{};
+  std::vector<uint64_t> _ids{};
   std::vector<std::string> _types{};
   std::vector<short> _isSubsetColl{};
 };
